Major update & preparation for Penilaain Coding 2

+++ app/src/main/java/com/example/grandatmahotel/ui/customer/act_detail/DetailReservasiViewModel.kt
@@ -26,7 +26,7 @@ class DetailReservasiViewModel(private val application: Application): ViewModel(
             val token = Utils.getToken(application)
             val response = ApiConfig.getApiService().getDetailReservasiCustomer(
                 token = "Bearer $token",
-                id = id
+                id = id.toLong()
             )
 
             val data = response.data
